requirements.txt
google-apitools==0.5.32
google-businessmessages==1.0.4
fasteners==0.15
httplib2==0.20.2
oauth2client==4.1.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
rsa==4.8
six==1.16.0